refactor(app): #6 renamed file and component to HeroSection
